# Optimize tag value parsing