```
goos: darwin
goarch: arm64
pkg: github.com/quickfixgo/quickfix
               │     old     │          new-no-indexbyte          │
               │   sec/op    │   sec/op     vs base               │
ParseMessage-8   695.2n ± 1%   628.2n ± 1%  -9.64% (p=0.000 n=20)
```